Fixed bug in get_group_coordinator
This fix addresses issue #64. The get_group_coordinator() method
returns ‘None’ for a zone which is in Stereo mode, when invoked on a
device which is a slave.

The method also needs to be invoked on a device that is initialized
with an IP address that belongs to the zone. This should not be the
case.

I’ve refactored the method to get the topology of all devices and
extract the coordinator for the given zone name. This allows us to get
the coordinator IP for any zone from any initialized SoCo object.