Added fields model and uid to method 'discover()'. Therefore, the ctor of SoCo has been changed.
model: presevers model name for later use
uid: get uid for a speaker without performing a deep scan (get_speaker_info())

In my project (shSonos on Github) the network is scanned periodically for new speakers. The 'uid' is the only unique identifier. If a new speaker was found, a deep scan is performed, otherwise there will be only an update after an elapsed period.