Ignore .# specified subclasses in Didl xml
Several music services seem to use a scheme for subclassing Didl classes by writing a subclass after the real class separated by .# e.g. object.container.album.musicAlbum.#AlbumView. This is not part of the Didl spec and so our implementation fails. This commit fixes this by stripping the sub-class, hence completely ignoring it. Fixes #399 and #388.