Add the ability to auto create vendor extended Didl classes
The extended classes are created as flat classes on top of the ones
the inherit from in the hierarchy. So if attributes are created or the
top level tag is different, then it will still be necessary to write
the class manually in data_structures.py. In all other cases it should
just work.